[setup][xs]: refactoring so data function is a module
