+++ test/data.js
@@ -1,0 +1,27 @@
+const path = require('path')
+const { spawn } = require('cross-spawn')
+
+const data = (...args) => {
+  return new Promise((resolve, reject) => {
+    const command = path.resolve(__dirname, '../bin/data.js')
+    const data = spawn(command, args)
+
+    let stdout = ''
+    data.stdout.on('data', data => {
+      stdout += data
+    })
+
+    data.on('error', err => {
+      reject(err)
+    })
+
+    data.on('close', code => {
+      resolve({
+        code,
+        stdout
+      })
+    })
+  })
+}
+
+module.exports.data = data
